rename: 지역 리스트 관련 컴포넌트 폴더 이동
