+++ src/features/search/ui/SearchRegionList/RecentSearchList/RecentSearchList.style.ts
