+++ src/pages/Search/ui/SearchPage.style.ts
@@ -1,5 +1,4 @@
 import styled from '@emotion/styled';
-import { List } from '@mui/material';
 
 const SearchPageWrap = styled.div`
   margin-top: 56px;
@@ -10,15 +9,7 @@ const InputWrap = styled.div`
   padding: 8px 16px;
 `;
 
-const RegionList = styled(List)`
-  padding: 0;
-`;
-
 export const S = {
   SearchPageWrap,
   InputWrap,
-};
-
-export const C = {
-  RegionList,
 };
